# Improve performance in `Bidiagonal` times `Diagonal`