This adds specialized methods to improve performance, and avoid allocations that were arising currently from the fallback tridiagonal implementations.

```julia
julia> using LinearAlgebra, BenchmarkTools

julia> n = 10000; B = Bidiagonal(rand(n), rand(n-1), :U); D = Diagonal(rand(size(B,1))); C = similar(B, size(B));

julia> @btime mul!($C, $B, $D);
  25.552 ms (3 allocations: 78.19 KiB) # v"1.12.0-DEV.870"
  25.559 ms (0 allocations: 0 bytes) # This PR

julia> C = similar(B);

julia> @btime mul!($C, $B, $D);
  23.551 μs (3 allocations: 78.19 KiB)  # v"1.12.0-DEV.870"
  7.123 μs (0 allocations: 0 bytes) # This PR, specialized method
```